Use SHRT_MAX instead INT_MAX as maximum tooltip width. This fixes problem with empty tooltips on High-DPI displays.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@27836 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseMerge/BaseView.cpp
@@ -2968,7 +2968,7 @@ INT_PTR CBaseView::OnToolHitTest(CPoint point, TOOLINFO* pTI) const
         // we want multi line tooltips
         CToolTipCtrl* pToolTip = AfxGetModuleThreadState()->m_pToolTip;
         if (pToolTip->GetSafeHwnd())
-            pToolTip->SetMaxTipWidth(INT_MAX);
+            pToolTip->SetMaxTipWidth(SHRT_MAX);
 
         return (textrect.PtInRect(point) ? 1 : 2);
     }
